repos/features_ingestion_pipeline/requirements.txt
attrs==21.2.0; python_version >= "3.7" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.7" and python_version < "4.0" and python_full_version >= "3.5.0"
aws-cdk.assets==1.126.0; python_version >= "3.6"
aws-cdk.aws-events==1.126.0; python_version >= "3.6"
aws-cdk.aws-iam==1.126.0; python_version >= "3.6"
aws-cdk.aws-kms==1.126.0; python_version >= "3.6"
aws-cdk.aws-s3-assets==1.126.0; python_version >= "3.6"
aws-cdk.aws-s3==1.126.0; python_version >= "3.6"
aws-cdk.aws-sagemaker==1.126.0; python_version >= "3.6"
aws-cdk.aws-ssm==1.126.0; python_version >= "3.6"
aws-cdk.cloud-assembly-schema==1.126.0; python_version >= "3.6"
aws-cdk.core==1.126.0; python_version >= "3.6"
aws-cdk.cx-api==1.126.0; python_version >= "3.6"
aws-cdk.region-info==1.126.0; python_version >= "3.6"
boto3==1.18.55; python_version >= "3.6"
botocore==1.21.55; python_version >= "3.6"
cattrs==1.8.0; python_version >= "3.7" and python_version < "4.0"
constructs==3.3.160; python_version >= "3.6"
dill==0.3.4; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.1.0"
google-pasta==0.2.0
importlib-metadata==4.8.1; python_version >= "3.6"
jmespath==0.10.0; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
jsii==1.36.0; python_version >= "3.6" and python_version < "4.0"
multiprocess==0.70.12.2
numpy==1.21.1; python_version >= "3.7" and python_full_version >= "3.7.1"
packaging==21.0; python_version >= "3.6"
pandas==1.3.3; python_full_version >= "3.7.1"
pathos==0.2.8
pox==0.3.0
ppft==1.6.6.4
protobuf3-to-dict==0.1.5
protobuf==3.18.1; python_version >= "3.5"
publication==0.0.3; python_version >= "3.6"
pyparsing==2.4.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6"
python-dateutil==2.8.2; python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.7.1" and (python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.6")
pytz==2021.3; python_full_version >= "3.7.1"
s3transfer==0.5.0; python_version >= "3.6"
sagemaker==2.59.7
six==1.16.0; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.3.0"
smdebug-rulesconfig==1.0.1; python_version >= "2.7"
typing-extensions==3.10.0.2; python_version >= "3.6" and python_version < "4.0"
urllib3==1.26.7; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6"
zipp==3.6.0; python_version >= "3.6"
